Revert "Moved codepen script to charting package"
This reverts commit 08d9582cd132cbd7b02d6f18c36f39f57b4b64b2.
